backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.97.1
python-multipart==0.0.20
pydantic==2.11.7
requests==2.31.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0